refactor calendar page into components while not breaking current functionality
