+++ frontend/src/components/calendar/CalendarSidebarContainer.tsx
@@ -1,0 +1,63 @@
+import React, { useCallback } from "react";
+import { useCalendarSidebarData } from "@/hooks/useCalendarSidebarData";
+import { CalendarSidebarPanel } from "@/components/calendar/CalendarSidebarPanel";
+
+type CalendarSidebarContainerProps = {
+  currentDate: Date;
+  onVisibleCalendarsChange: (calendarIds: string[] | undefined) => void;
+  onDateChange: (next: Date) => void;
+  selectedCalendarId: string | null;
+};
+
+export function CalendarSidebarContainer({
+  currentDate,
+  onVisibleCalendarsChange,
+  onDateChange,
+  selectedCalendarId,
+}: CalendarSidebarContainerProps) {
+  const { myCalendars, otherCalendars, isLoading, error } = useCalendarSidebarData();
+
+  const handleCalendarItemClick = useCallback(
+    (calendarId: string) => {
+      if (selectedCalendarId === calendarId) {
+        onVisibleCalendarsChange(undefined);
+        return;
+      }
+      onVisibleCalendarsChange([calendarId]);
+    },
+    [onVisibleCalendarsChange, selectedCalendarId],
+  );
+
+  React.useEffect(() => {
+    if (!selectedCalendarId || isLoading || error) {
+      return;
+    }
+    const allItems = [...myCalendars, ...otherCalendars];
+    const hasSelectedCalendar = allItems.some(
+      (item) => item.calendarId === selectedCalendarId,
+    );
+    if (!hasSelectedCalendar) {
+      onVisibleCalendarsChange(undefined);
+    }
+  }, [
+    selectedCalendarId,
+    isLoading,
+    error,
+    myCalendars,
+    otherCalendars,
+    onVisibleCalendarsChange,
+  ]);
+
+  return (
+    <CalendarSidebarPanel
+      currentDate={currentDate}
+      onDateChange={onDateChange}
+      selectedCalendarId={selectedCalendarId}
+      myCalendars={myCalendars}
+      otherCalendars={otherCalendars}
+      isLoading={isLoading}
+      error={error}
+      onCalendarItemClick={handleCalendarItemClick}
+    />
+  );
+}
