+++ frontend/src/components/calendar/CalendarViews.tsx
@@ -1,0 +1,923 @@
+import React, { useMemo } from "react";
+import {
+  addDays,
+  endOfMonth,
+  endOfYear,
+  format,
+  isSameDay,
+  isSameMonth,
+  startOfDay,
+  startOfMonth,
+  startOfWeek,
+  startOfYear,
+} from "date-fns";
+import type { CalendarDTO, EventDTO, CalendarViewType } from "@/lib/api/calendarApi";
+import type { EventPanelPosition } from "@/components/calendar/types";
+import { computePanelPosition } from "@/components/calendar/utils";
+
+export type WeekViewProps = {
+  currentDate: Date;
+  events: EventDTO[];
+  calendars: CalendarDTO[];
+  isLoading: boolean;
+  error: Error | null;
+  onTimeSlotClick: (start: Date, position: EventPanelPosition) => void;
+  onEventClick: (event: EventDTO, position: EventPanelPosition) => void;
+  onEventTimeChange: (event: EventDTO, start: Date, end: Date) => Promise<void>;
+};
+
+export function WeekView({
+  currentDate,
+  events,
+  calendars,
+  isLoading,
+  error,
+  onTimeSlotClick,
+  onEventClick,
+  onEventTimeChange,
+}: WeekViewProps) {
+  const start = startOfWeek(currentDate, { weekStartsOn: 1 });
+  const days = useMemo(
+    () => Array.from({ length: 7 }, (_, index) => addDays(start, index)),
+    [start],
+  );
+  const hours = useMemo(
+    () => Array.from({ length: 24 }, (_, index) => index),
+    [],
+  );
+
+  const [dragState, setDragState] = React.useState<{
+    eventId: string;
+    mode: "move" | "resize";
+    originY: number;
+    originX: number;
+    originalStart: Date;
+    originalEnd: Date;
+  } | null>(null);
+  const [previewTimes, setPreviewTimes] = React.useState<
+    Record<string, { start: Date; end: Date }>
+  >({});
+  const [suppressClick, setSuppressClick] = React.useState(false);
+
+  const calendarColorById = useMemo(() => {
+    const map = new Map<string, string>();
+    calendars.forEach((cal) => map.set(cal.id, cal.color));
+    return map;
+  }, [calendars]);
+  const eventById = useMemo(() => {
+    const map = new Map<string, EventDTO>();
+    events.forEach((ev) => map.set(ev.id, ev));
+    return map;
+  }, [events]);
+
+  const handleMouseMove: React.MouseEventHandler<HTMLDivElement> = (e) => {
+    if (!dragState) return;
+    if ((e.buttons & 1) === 0) return;
+
+    const pixelsPerMinute = 48 / 60;
+    const stepMinutes = 30;
+    const deltaY = e.clientY - dragState.originY;
+    const rawMinutes = deltaY / pixelsPerMinute;
+    const snappedMinutes = Math.round(rawMinutes / stepMinutes) * stepMinutes;
+
+    let dayOffset = 0;
+    if (dragState.mode === "move") {
+      const gridElement = e.currentTarget as HTMLDivElement;
+      const rect = gridElement.getBoundingClientRect();
+      const totalWidth = rect.width - 60;
+      const dayWidth = totalWidth / 7;
+      const deltaX = e.clientX - dragState.originX;
+      dayOffset = Math.round(deltaX / dayWidth);
+    }
+
+    if (snappedMinutes !== 0 || dayOffset !== 0) {
+      setSuppressClick(true);
+    }
+
+    const totalMinutesOffset = snappedMinutes + dayOffset * 24 * 60;
+    const newStart = new Date(
+      dragState.originalStart.getTime() + totalMinutesOffset * 60000,
+    );
+    let newEnd: Date;
+
+    if (dragState.mode === "move") {
+      newEnd = new Date(
+        dragState.originalEnd.getTime() + totalMinutesOffset * 60000,
+      );
+    } else {
+      const minDurationMinutes = 15;
+      const candidateEnd = new Date(
+        dragState.originalEnd.getTime() + snappedMinutes * 60000,
+      );
+      if (
+        candidateEnd.getTime() - dragState.originalStart.getTime() <
+        minDurationMinutes * 60000
+      ) {
+        newEnd = new Date(
+          dragState.originalStart.getTime() + minDurationMinutes * 60000,
+        );
+      } else {
+        newEnd = candidateEnd;
+      }
+    }
+
+    setPreviewTimes((prev) => ({
+      ...prev,
+      [dragState.eventId]: { start: newStart, end: newEnd },
+    }));
+  };
+
+  const finishDrag = async () => {
+    if (!dragState) return;
+    const preview = previewTimes[dragState.eventId];
+    const baseEvent = eventById.get(dragState.eventId);
+    setDragState(null);
+    setPreviewTimes((prev) => {
+      const next = { ...prev };
+      delete next[dragState.eventId];
+      return next;
+    });
+    if (!preview || !baseEvent) return;
+    await onEventTimeChange(baseEvent, preview.start, preview.end);
+  };
+
+  return (
+    <div className="flex h-full flex-col rounded-3xl bg-white shadow-sm">
+      <div className="grid grid-cols-[60px_repeat(7,minmax(0,1fr))] border-b bg-white text-xs font-medium text-gray-500">
+        <div className="border-r px-2 py-2" />
+        {days.map((day) => (
+          <div
+            key={day.toISOString()}
+            className="flex flex-col items-center border-r px-2 py-2 last:border-r-0"
+          >
+            <span>{format(day, "EEE")}</span>
+            <span className="mt-1 rounded-full px-1.5 py-0.5 text-sm font-semibold text-gray-800">
+              {format(day, "d")}
+            </span>
+          </div>
+        ))}
+      </div>
+
+      <div
+        className="grid flex-1 grid-cols-[60px_repeat(7,minmax(0,1fr))] overflow-auto text-xs"
+        onMouseMove={handleMouseMove}
+        onMouseUp={() => void finishDrag()}
+        onMouseLeave={() => void finishDrag()}
+      >
+        <div className="border-r bg-white">
+          {hours.map((hour) => (
+            <div
+              key={hour}
+              className="h-12 border-b border-gray-100 px-2 py-1 text-right text-[11px] text-gray-400"
+            >
+              {format(new Date().setHours(hour, 0, 0, 0), "ha")}
+            </div>
+          ))}
+        </div>
+
+        {days.map((day) => {
+          const dayStart = startOfDay(day);
+          const dayEnd = addDays(dayStart, 1);
+          const dayEvents = events.filter((event) => {
+            const evStart = new Date(event.start_datetime);
+            const evEnd = new Date(event.end_datetime);
+            return evEnd > dayStart && evStart < dayEnd;
+          });
+
+          return (
+            <div key={day.toISOString()} className="relative border-r last:border-r-0">
+              {hours.map((hour) => {
+                const slotStart = new Date(dayStart);
+                slotStart.setHours(hour, 0, 0, 0);
+                return (
+                  <button
+                    key={hour}
+                    type="button"
+                    className="h-12 w-full border-b border-gray-100 bg-white text-left hover:bg-blue-50"
+                    onClick={(e) => {
+                      const rect = e.currentTarget.getBoundingClientRect();
+                      const position = computePanelPosition(rect);
+                      onTimeSlotClick(slotStart, position);
+                    }}
+                  />
+                );
+              })}
+
+              {dayEvents.map((event) => {
+                const preview = previewTimes[event.id];
+                const evStart = preview ? preview.start : new Date(event.start_datetime);
+                const evEnd = preview ? preview.end : new Date(event.end_datetime);
+                const clampedStart = evStart < dayStart ? dayStart : evStart;
+                const clampedEnd = evEnd > dayEnd ? dayEnd : evEnd;
+                const startMinutes =
+                  (clampedStart.getTime() - dayStart.getTime()) / 60000;
+                const durationMinutes =
+                  (clampedEnd.getTime() - clampedStart.getTime()) / 60000 || 30;
+                const topPercent = (startMinutes / (24 * 60)) * 100;
+                const heightPercent = (durationMinutes / (24 * 60)) * 100;
+                const backgroundColor =
+                  event.color || calendarColorById.get(event.calendar_id || "") || "#1E88E5";
+
+                return (
+                  <button
+                    key={event.id + event.start_datetime}
+                    type="button"
+                    onClick={(e) => {
+                      if (suppressClick) {
+                        e.preventDefault();
+                        setSuppressClick(false);
+                        return;
+                      }
+                      const rect = e.currentTarget.getBoundingClientRect();
+                      const position = computePanelPosition(rect);
+                      onEventClick(event, position);
+                    }}
+                    onMouseDown={(e) => {
+                      if (event.is_recurring || e.button !== 0) return;
+                      e.preventDefault();
+                      e.stopPropagation();
+                      setDragState({
+                        eventId: event.id,
+                        mode: "move",
+                        originY: e.clientY,
+                        originX: e.clientX,
+                        originalStart: new Date(event.start_datetime),
+                        originalEnd: new Date(event.end_datetime),
+                      });
+                    }}
+                    className="absolute left-1 right-1 rounded-md px-1.5 py-0.5 text-[11px] text-white shadow-sm"
+                    style={{ top: `${topPercent}%`, height: `${heightPercent}%`, backgroundColor }}
+                  >
+                    <div className="truncate font-semibold">{event.title}</div>
+                    <div className="truncate opacity-90">
+                      {format(evStart, "HH:mm")} - {format(evEnd, "HH:mm")}
+                    </div>
+                    {!event.is_recurring && (
+                      <div
+                        className="mt-1 h-1 w-full cursor-row-resize rounded bg-white/60"
+                        onMouseDown={(e) => {
+                          if (e.button !== 0) return;
+                          e.preventDefault();
+                          e.stopPropagation();
+                          setDragState({
+                            eventId: event.id,
+                            mode: "resize",
+                            originY: e.clientY,
+                            originX: e.clientX,
+                            originalStart: new Date(event.start_datetime),
+                            originalEnd: new Date(event.end_datetime),
+                          });
+                        }}
+                      />
+                    )}
+                  </button>
+                );
+              })}
+
+              {isLoading && <div className="pointer-events-none absolute inset-0 bg-white/40" />}
+              {error && (
+                <div className="pointer-events-none absolute inset-x-2 top-2 rounded bg-red-50 px-2 py-1 text-[10px] text-red-600">
+                  Failed to load events.
+                </div>
+              )}
+            </div>
+          );
+        })}
+      </div>
+    </div>
+  );
+}
+
+export type DayViewProps = WeekViewProps;
+
+export function DayView({
+  currentDate,
+  events,
+  calendars,
+  isLoading,
+  error,
+  onTimeSlotClick,
+  onEventClick,
+  onEventTimeChange,
+}: DayViewProps) {
+  const dayStart = startOfDay(currentDate);
+  const dayEnd = addDays(dayStart, 1);
+  const hours = useMemo(() => Array.from({ length: 24 }, (_, index) => index), []);
+  const [dragState, setDragState] = React.useState<{
+    eventId: string;
+    mode: "move" | "resize";
+    originY: number;
+    originalStart: Date;
+    originalEnd: Date;
+  } | null>(null);
+  const [previewTimes, setPreviewTimes] = React.useState<
+    Record<string, { start: Date; end: Date }>
+  >({});
+  const [suppressClick, setSuppressClick] = React.useState(false);
+
+  const dayEvents = events.filter((event) => {
+    const evStart = new Date(event.start_datetime);
+    const evEnd = new Date(event.end_datetime);
+    return evEnd > dayStart && evStart < dayEnd;
+  });
+  const calendarColorById = useMemo(() => {
+    const map = new Map<string, string>();
+    calendars.forEach((cal) => map.set(cal.id, cal.color));
+    return map;
+  }, [calendars]);
+  const eventById = useMemo(() => {
+    const map = new Map<string, EventDTO>();
+    events.forEach((ev) => map.set(ev.id, ev));
+    return map;
+  }, [events]);
+
+  const handleMouseMove: React.MouseEventHandler<HTMLDivElement> = (e) => {
+    if (!dragState) return;
+    if ((e.buttons & 1) === 0) return;
+    const pixelsPerMinute = 48 / 60;
+    const stepMinutes = 30;
+    const deltaY = e.clientY - dragState.originY;
+    const snappedMinutes = Math.round((deltaY / pixelsPerMinute) / stepMinutes) * stepMinutes;
+    if (snappedMinutes !== 0) setSuppressClick(true);
+
+    const newStart = new Date(dragState.originalStart.getTime() + snappedMinutes * 60000);
+    let newEnd: Date;
+    if (dragState.mode === "move") {
+      newEnd = new Date(dragState.originalEnd.getTime() + snappedMinutes * 60000);
+    } else {
+      const minDurationMinutes = 15;
+      const candidateEnd = new Date(dragState.originalEnd.getTime() + snappedMinutes * 60000);
+      if (candidateEnd.getTime() - dragState.originalStart.getTime() < minDurationMinutes * 60000) {
+        newEnd = new Date(dragState.originalStart.getTime() + minDurationMinutes * 60000);
+      } else {
+        newEnd = candidateEnd;
+      }
+    }
+    setPreviewTimes((prev) => ({ ...prev, [dragState.eventId]: { start: newStart, end: newEnd } }));
+  };
+
+  const finishDrag = async () => {
+    if (!dragState) return;
+    const preview = previewTimes[dragState.eventId];
+    const baseEvent = eventById.get(dragState.eventId);
+    setDragState(null);
+    setPreviewTimes((prev) => {
+      const next = { ...prev };
+      delete next[dragState.eventId];
+      return next;
+    });
+    if (!preview || !baseEvent) return;
+    await onEventTimeChange(baseEvent, preview.start, preview.end);
+  };
+
+  return (
+    <div className="flex flex-col rounded-xl bg-white">
+      <div className="grid grid-cols-[60px_minmax(0,1fr)] border-b bg-white text-xs font-medium text-gray-500">
+        <div className="border-r px-2 py-2" />
+        <div className="flex flex-col px-2 py-2">
+          <span>{format(currentDate, "EEE")}</span>
+          <span className="mt-1 rounded-full px-1.5 py-0.5 text-sm font-semibold text-gray-800">
+            {format(currentDate, "d")}
+          </span>
+        </div>
+      </div>
+      <div
+        className="grid flex-1 grid-cols-[60px_minmax(0,1fr)] overflow-auto text-xs"
+        onMouseMove={handleMouseMove}
+        onMouseUp={() => void finishDrag()}
+        onMouseLeave={() => void finishDrag()}
+      >
+        <div className="border-r bg-white">
+          {hours.map((hour) => (
+            <div
+              key={hour}
+              className="inline-block h-12 border-b border-gray-100 px-2 text-right text-[11px] text-gray-400"
+            >
+              {format(new Date().setHours(hour, 0, 0, 0), "ha")}
+            </div>
+          ))}
+        </div>
+        <div className="relative">
+          {hours.map((hour) => {
+            const slotStart = new Date(dayStart);
+            slotStart.setHours(hour, 0, 0, 0);
+            return (
+              <button
+                key={hour}
+                type="button"
+                className="h-12 w-full border-b border-gray-100 bg-white text-left hover:bg-blue-50"
+                onClick={(e) => {
+                  const rect = e.currentTarget.getBoundingClientRect();
+                  const position = computePanelPosition(rect, "day");
+                  onTimeSlotClick(slotStart, position);
+                }}
+              />
+            );
+          })}
+
+          {dayEvents.map((event) => {
+            const preview = previewTimes[event.id];
+            const evStart = preview ? preview.start : new Date(event.start_datetime);
+            const evEnd = preview ? preview.end : new Date(event.end_datetime);
+            const clampedStart = evStart < dayStart ? dayStart : evStart;
+            const clampedEnd = evEnd > dayEnd ? dayEnd : evEnd;
+            const startMinutes = (clampedStart.getTime() - dayStart.getTime()) / 60000;
+            const durationMinutes = (clampedEnd.getTime() - clampedStart.getTime()) / 60000 || 30;
+            const topPercent = (startMinutes / (24 * 60)) * 100;
+            const heightPercent = (durationMinutes / (24 * 60)) * 100;
+            const backgroundColor =
+              event.color || calendarColorById.get(event.calendar_id || "") || "#1E88E5";
+
+            return (
+              <button
+                key={event.id + event.start_datetime}
+                type="button"
+                onClick={(e) => {
+                  if (suppressClick) {
+                    e.preventDefault();
+                    setSuppressClick(false);
+                    return;
+                  }
+                  onEventClick(event, computePanelPosition(null, "day"));
+                }}
+                onMouseDown={(e) => {
+                  if (event.is_recurring || e.button !== 0) return;
+                  e.preventDefault();
+                  e.stopPropagation();
+                  setDragState({
+                    eventId: event.id,
+                    mode: "move",
+                    originY: e.clientY,
+                    originalStart: new Date(event.start_datetime),
+                    originalEnd: new Date(event.end_datetime),
+                  });
+                }}
+                className="absolute left-1 right-1 rounded-md px-1.5 py-0.5 text-[11px] text-white shadow-sm"
+                style={{ top: `${topPercent}%`, height: `${heightPercent}%`, backgroundColor }}
+              >
+                <div className="truncate font-semibold">{event.title}</div>
+                <div className="truncate opacity-90">
+                  {format(evStart, "HH:mm")} - {format(evEnd, "HH:mm")}
+                </div>
+                {!event.is_recurring && (
+                  <div
+                    className="mt-1 h-1 w-full cursor-row-resize rounded bg-white/60"
+                    onMouseDown={(e) => {
+                      if (e.button !== 0) return;
+                      e.preventDefault();
+                      e.stopPropagation();
+                      setDragState({
+                        eventId: event.id,
+                        mode: "resize",
+                        originY: e.clientY,
+                        originalStart: new Date(event.start_datetime),
+                        originalEnd: new Date(event.end_datetime),
+                      });
+                    }}
+                  />
+                )}
+              </button>
+            );
+          })}
+
+          {isLoading && <div className="pointer-events-none absolute inset-0 bg-white/40" />}
+          {error && (
+            <div className="pointer-events-none absolute inset-x-2 top-2 rounded bg-red-50 px-2 py-1 text-[10px] text-red-600">
+              Failed to load events.
+            </div>
+          )}
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export type MonthViewProps = {
+  currentDate: Date;
+  events: EventDTO[];
+  calendars: CalendarDTO[];
+  isLoading: boolean;
+  error: Error | null;
+  onDaySelect: (day: Date) => void;
+  onEventClick: (event: EventDTO) => void;
+};
+
+export function MonthView({
+  currentDate,
+  events,
+  calendars,
+  isLoading,
+  error,
+  onDaySelect,
+  onEventClick,
+}: MonthViewProps) {
+  const startMonth = startOfMonth(currentDate);
+  const gridStart = startOfWeek(startMonth, { weekStartsOn: 1 });
+  const days = useMemo(
+    () => Array.from({ length: 42 }, (_, index) => addDays(gridStart, index)),
+    [gridStart],
+  );
+  const calendarColorById = useMemo(() => {
+    const map = new Map<string, string>();
+    calendars.forEach((cal) => map.set(cal.id, cal.color));
+    return map;
+  }, [calendars]);
+  const eventsByDay = useMemo(() => {
+    const result = new Map<string, EventDTO[]>();
+    days.forEach((day) => result.set(format(day, "yyyy-MM-dd"), []));
+    events.forEach((event) => {
+      const evStart = new Date(event.start_datetime);
+      const evEnd = new Date(event.end_datetime);
+      days.forEach((day) => {
+        const dayStart = startOfDay(day);
+        const dayEnd = addDays(dayStart, 1);
+        if (evEnd > dayStart && evStart < dayEnd) {
+          result.get(format(day, "yyyy-MM-dd"))?.push(event);
+        }
+      });
+    });
+    return result;
+  }, [days, events]);
+
+  const weekdayLabels = ["M", "T", "W", "T", "F", "S", "S"];
+  const today = new Date();
+
+  return (
+    <div className="flex h-full flex-col rounded-xl bg-white">
+      <div className="flex items-center justify-between px-4 py-2 text-sm font-semibold text-gray-700">
+        {isLoading && <span className="text-xs text-gray-400">Loading…</span>}
+        {error && <span className="text-xs text-red-500">Failed to load events.</span>}
+      </div>
+      <div className="grid flex-1 grid-rows-[auto_1fr]">
+        <div className="grid grid-cols-7 bg-white text-[11px] font-medium text-gray-500">
+          {weekdayLabels.map((label) => (
+            <div key={label} className="flex items-center justify-center py-1">
+              {label}
+            </div>
+          ))}
+        </div>
+        <div className="grid flex-1 grid-cols-7 grid-rows-6 text-xs">
+          {days.map((day) => {
+            const inMonth = isSameMonth(day, startMonth);
+            const isSelected = isSameDay(day, currentDate);
+            const isToday = isSameDay(day, today);
+            const dayEvents = eventsByDay.get(format(day, "yyyy-MM-dd")) || [];
+            return (
+              <button
+                key={day.toISOString()}
+                type="button"
+                className="flex h-full flex-col border-b border-r bg-white px-1.5 py-1"
+                onClick={() => onDaySelect(day)}
+              >
+                <div className="mb-1 flex items-center justify-center text-[11px]">
+                  <span
+                    className={`inline-flex h-6 w-6 items-center justify-center rounded-full ${
+                      isSelected
+                        ? "bg-blue-600 text-white"
+                        : isToday
+                        ? "border border-blue-500 text-blue-700"
+                        : inMonth
+                        ? "text-gray-800"
+                        : "text-gray-300"
+                    }`}
+                  >
+                    {format(day, "d")}
+                  </span>
+                </div>
+                <div className="space-y-0.5">
+                  {dayEvents.slice(0, 3).map((event) => {
+                    const color =
+                      event.color || calendarColorById.get(event.calendar_id || "") || "#1E88E5";
+                    return (
+                      <div
+                        key={event.id + event.start_datetime}
+                        className="flex cursor-pointer items-center gap-1 truncate rounded px-1 py-0.5 text-[11px] text-gray-900 hover:bg-gray-100"
+                        onClick={(e) => {
+                          e.stopPropagation();
+                          onEventClick(event);
+                        }}
+                      >
+                        <span className="h-2 w-2 rounded-full" style={{ backgroundColor: color }} />
+                        <span className="truncate">{event.title}</span>
+                      </div>
+                    );
+                  })}
+                  {dayEvents.length > 3 && (
+                    <div className="text-[10px] text-gray-500">+{dayEvents.length - 3} more</div>
+                  )}
+                </div>
+              </button>
+            );
+          })}
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export type AgendaViewProps = {
+  currentDate: Date;
+  events: EventDTO[];
+  calendars: CalendarDTO[];
+  isLoading: boolean;
+  error: Error | null;
+  onEventClick: (event: EventDTO, position: EventPanelPosition) => void;
+};
+
+export function AgendaView({
+  currentDate,
+  events,
+  calendars,
+  isLoading,
+  error,
+  onEventClick,
+}: AgendaViewProps) {
+  const calendarColorById = useMemo(() => {
+    const map = new Map<string, string>();
+    calendars.forEach((cal) => map.set(cal.id, cal.color));
+    return map;
+  }, [calendars]);
+  const eventsByDate = useMemo(() => {
+    const grouped = new Map<string, EventDTO[]>();
+    [...events]
+      .sort(
+        (a, b) =>
+          new Date(a.start_datetime).getTime() - new Date(b.start_datetime).getTime(),
+      )
+      .forEach((event) => {
+        const key = format(new Date(event.start_datetime), "yyyy-MM-dd");
+        const bucket = grouped.get(key);
+        if (bucket) {
+          bucket.push(event);
+        } else {
+          grouped.set(key, [event]);
+        }
+      });
+    return grouped;
+  }, [events]);
+  const dateKeys = Array.from(eventsByDate.keys());
+
+  return (
+    <div className="flex h-full flex-col rounded-xl bg-white shadow-sm">
+      <div className="flex items-center justify-between border-b px-4 py-2 text-sm font-semibold text-gray-700">
+        <span>
+          {format(currentDate, "MMM d, yyyy")} – {format(addDays(currentDate, 7), "MMM d, yyyy")}
+        </span>
+        {isLoading && <span className="text-xs text-gray-400">Loading…</span>}
+        {error && <span className="text-xs text-red-500">Failed to load events.</span>}
+      </div>
+      <div className="flex-1 overflow-auto">
+        {dateKeys.length === 0 && !isLoading && !error && (
+          <div className="flex h-full items-center justify-center text-sm text-gray-500">
+            No events in this range.
+          </div>
+        )}
+        {dateKeys.map((key) => {
+          const day = new Date(key);
+          const dayEvents = eventsByDate.get(key) || [];
+          return (
+            <div key={key} className="border-b px-4 py-3 text-sm">
+              <div className="mb-2 font-semibold text-gray-800">{format(day, "EEE, MMM d")}</div>
+              <ul className="space-y-1">
+                {dayEvents.map((event) => {
+                  const color =
+                    event.color || calendarColorById.get(event.calendar_id || "") || "#1E88E5";
+                  const start = new Date(event.start_datetime);
+                  const end = new Date(event.end_datetime);
+                  return (
+                    <li key={event.id + event.start_datetime}>
+                      <button
+                        type="button"
+                        onClick={(e) => {
+                          const rect = e.currentTarget.getBoundingClientRect();
+                          const position = computePanelPosition(rect);
+                          onEventClick(event, position);
+                        }}
+                        className="flex w-full items-center justify-between rounded px-2 py-1 text-left hover:bg-gray-50"
+                      >
+                        <div className="flex items-center gap-2">
+                          <span className="h-2 w-2 rounded-full" style={{ backgroundColor: color }} />
+                          <span className="text-xs text-gray-500">
+                            {format(start, "HH:mm")} – {format(end, "HH:mm")}
+                          </span>
+                          <span className="truncate text-sm text-gray-900">{event.title}</span>
+                        </div>
+                      </button>
+                    </li>
+                  );
+                })}
+              </ul>
+            </div>
+          );
+        })}
+      </div>
+    </div>
+  );
+}
+
+export type YearViewProps = {
+  currentDate: Date;
+  onDaySelect: (day: Date) => void;
+};
+
+export function YearView({ currentDate, onDaySelect }: YearViewProps) {
+  const yearStart = startOfYear(currentDate);
+  const yearEnd = endOfYear(currentDate);
+  const months = useMemo(() => {
+    void yearEnd;
+    return Array.from({ length: 12 }, (_, index) => {
+      const d = new Date(yearStart);
+      d.setMonth(index, 1);
+      return d;
+    });
+  }, [yearStart, yearEnd]);
+  const today = new Date();
+
+  return (
+    <div className="flex h-full flex-col rounded-xl bg-white">
+      <div className="grid flex-1 grid-cols-3 gap-4 overflow-auto p-4 text-[11px]">
+        {months.map((monthDate) => {
+          const startMonth = startOfMonth(monthDate);
+          const gridStart = startOfWeek(startMonth, { weekStartsOn: 1 });
+          const days = Array.from({ length: 42 }, (_, index) => addDays(gridStart, index));
+          const weekdayLabels = ["M", "T", "W", "T", "F", "S", "S"];
+
+          return (
+            <div key={monthDate.toISOString()} className="rounded bg-white p-2">
+              <div className="mb-1 text-center text-[11px] font-semibold text-gray-700">
+                {format(monthDate, "MMMM")}
+              </div>
+              <div className="grid grid-cols-7 place-items-center gap-0.5 text-[10px] text-gray-400">
+                {weekdayLabels.map((label) => (
+                  <div key={label} className="flex h-4 items-center justify-center">
+                    {label}
+                  </div>
+                ))}
+                {days.map((day) => {
+                  const inMonth = isSameMonth(day, startMonth);
+                  const isToday = isSameDay(day, today);
+                  const className = `flex h-5 w-5 items-center justify-center rounded-full ${
+                    !inMonth
+                      ? "text-gray-300"
+                      : isToday
+                      ? "border border-blue-500 bg-white text-blue-700"
+                      : "text-gray-700 hover:bg-white"
+                  }`;
+                  return (
+                    <button
+                      key={day.toISOString()}
+                      type="button"
+                      className={className}
+                      onClick={() => onDaySelect(day)}
+                    >
+                      {format(day, "d")}
+                    </button>
+                  );
+                })}
+              </div>
+            </div>
+          );
+        })}
+      </div>
+    </div>
+  );
+}
+
+export type MiniMonthCalendarProps = {
+  currentDate: Date;
+  onDateChange: (next: Date) => void;
+};
+
+export function MiniMonthCalendar({ currentDate, onDateChange }: MiniMonthCalendarProps) {
+  const startMonth = startOfMonth(currentDate);
+  const gridStart = startOfWeek(startMonth, { weekStartsOn: 1 });
+  const days = useMemo(
+    () => Array.from({ length: 42 }, (_, index) => addDays(gridStart, index)),
+    [gridStart],
+  );
+  const weekdayLabels = ["M", "T", "W", "T", "F", "S", "S"];
+  const today = new Date();
+
+  return (
+    <div className="mb-6 rounded-xl p-3">
+      <div className="mb-2 flex items-center justify-between text-xs font-semibold text-gray-700">
+        <span>{format(currentDate, "MMMM yyyy")}</span>
+      </div>
+      <div className="grid grid-cols-7 gap-1 text-[11px] text-gray-500">
+        {weekdayLabels.map((label) => (
+          <div key={label} className="flex h-5 items-center justify-center">
+            {label}
+          </div>
+        ))}
+        {days.map((day) => {
+          const inMonth = isSameMonth(day, startMonth);
+          const isSelected = isSameDay(day, currentDate);
+          const isToday = isSameDay(day, today);
+          const className = `flex h-7 w-7 items-center justify-center rounded-full text-xs ${
+            isSelected
+              ? "bg-blue-600 text-white"
+              : isToday
+              ? "border border-blue-500 text-blue-700"
+              : !inMonth
+              ? "text-gray-300"
+              : "text-gray-700 hover:bg-white"
+          }`;
+          return (
+            <button
+              key={day.toISOString()}
+              type="button"
+              className={className}
+              onClick={() => onDateChange(day)}
+            >
+              {format(day, "d")}
+            </button>
+          );
+        })}
+      </div>
+    </div>
+  );
+}
+
+type CalendarViewRouterProps = {
+  currentView: CalendarViewType;
+  currentDate: Date;
+  events: EventDTO[];
+  calendars: CalendarDTO[];
+  isLoading: boolean;
+  error: Error | null;
+  onTimeSlotClick: (start: Date, position: EventPanelPosition) => void;
+  onEventClick: (event: EventDTO, position: EventPanelPosition) => void;
+  onEventTimeChange: (event: EventDTO, start: Date, end: Date) => Promise<void>;
+  onDaySelect: (day: Date) => void;
+};
+
+export function CalendarViewRouter({
+  currentView,
+  currentDate,
+  events,
+  calendars,
+  isLoading,
+  error,
+  onTimeSlotClick,
+  onEventClick,
+  onEventTimeChange,
+  onDaySelect,
+}: CalendarViewRouterProps) {
+  if (currentView === "week") {
+    return (
+      <WeekView
+        currentDate={currentDate}
+        events={events}
+        calendars={calendars}
+        isLoading={isLoading}
+        error={error}
+        onTimeSlotClick={onTimeSlotClick}
+        onEventClick={onEventClick}
+        onEventTimeChange={onEventTimeChange}
+      />
+    );
+  }
+  if (currentView === "day") {
+    return (
+      <DayView
+        currentDate={currentDate}
+        events={events}
+        calendars={calendars}
+        isLoading={isLoading}
+        error={error}
+        onTimeSlotClick={onTimeSlotClick}
+        onEventClick={onEventClick}
+        onEventTimeChange={onEventTimeChange}
+      />
+    );
+  }
+  if (currentView === "month") {
+    return (
+      <MonthView
+        currentDate={currentDate}
+        events={events}
+        calendars={calendars}
+        isLoading={isLoading}
+        error={error}
+        onDaySelect={onDaySelect}
+        onEventClick={(event) => onEventClick(event, { top: 120, left: 320 })}
+      />
+    );
+  }
+  if (currentView === "agenda") {
+    return (
+      <AgendaView
+        currentDate={currentDate}
+        events={events}
+        calendars={calendars}
+        isLoading={isLoading}
+        error={error}
+        onEventClick={onEventClick}
+      />
+    );
+  }
+  if (currentView === "year") {
+    return <YearView currentDate={currentDate} onDaySelect={onDaySelect} />;
+  }
+  return null;
+}
