+++ frontend/src/components/calendar/EventDialogContainer.tsx
@@ -1,0 +1,22 @@
+import type { CalendarDTO, EventDTO } from "@/lib/api/calendarApi";
+import type { CalendarDialogMode, EventPanelPosition } from "@/components/calendar/types";
+import { EventPanelDialog } from "@/components/calendar/EventPanelDialog";
+
+type EventDialogContainerProps = {
+  open: boolean;
+  mode: CalendarDialogMode;
+  onModeChange: (mode: CalendarDialogMode) => void;
+  onOpenChange: (open: boolean) => void;
+  start: Date | null;
+  end: Date | null;
+  event: EventDTO | null;
+  calendars: CalendarDTO[];
+  preferredCalendarId?: string | null;
+  onSave: (payload: { action: () => Promise<void> }) => Promise<void>;
+  onDelete?: (event: EventDTO) => Promise<void>;
+  position: EventPanelPosition | null;
+};
+
+export function EventDialogContainer(props: EventDialogContainerProps) {
+  return <EventPanelDialog {...props} />;
+}
